requirements.txt
python-dotenv==1.2.2
pytz==2023.3.post1
Requests==2.33.1
tabulate==0.9.0
